backend/requirements.txt
Django==3.0.3
Pillow==7.1.2
djangorestframework==3.11.0
markdown==3.2.2
django-filter==2.3.0
psycopg2-binary==2.8.5
django-cors-headers==3.4.0
djoser==2.0.3
djangorestframework_simplejwt==4.4.0

channels==2.4.0
channels-redis==2.4.2



